+++ includes/rcon.sh
@@ -1,57 +1,54 @@
 # shellcheck disable=SC2148
 
+function get_time() {
+    date '+[%H:%M:%S]'
+}
+
 function save_and_shutdown_server() {
-    rconcli 'broadcast Server-shutdown-was-requested-init-saving'
-    rconcli 'save'
-    rconcli 'broadcast Done-saving-server-shuts-down-now'
+    rconcli broadcast "$(get_time) Server shutdown requested. Saving..."
+    rconcli save
+    rconcli broadcast "$(get_time) Saving done. Server shutting down..."
 }
 
 function broadcast_automatic_restart() {
-    time=$(date '+%H:%M:%S')
-
     for ((counter=1; counter<=15; counter++)); do
-        rconcli "broadcast ${time}-AUTOMATIC-RESTART-IN-$counter-MINUTES"
+        rconcli "broadcast $(get_time)-AUTOMATIC-RESTART-IN-$counter-MINUTES"
         sleep 1
     done
-    rconcli 'broadcast Saving-world-before-restart...'
-    rconcli 'save'
-    rconcli 'broadcast Saving-done'
-    rconcli 'broadcast Creating-backup'
+    rconcli broadcast "$(get_time) Saving world before restart..."
+    rconcli save
+    rconcli broadcast "$(get_time) Saving done"
+    rconcli broadcast "$(get_time) Creating backup..."
     rcon "Shutdown 10"
 }
 
 function broadcast_backup_start() {
-    time=$(date '+%H:%M:%S')
-
-    rconcli "broadcast ${time}-Saving-in-5-seconds"
+    rconcli broadcast "$(get_time) Saving in 5 seconds..."
     sleep 5
-    rconcli 'broadcast Saving-world...'
-    rconcli 'save'
-    rconcli 'broadcast Saving-done'
-    rconcli 'broadcast Creating-backup'
+    rconcli broadcast "$(get_time) Saving world..."
+    rconcli save
+    rconcli broadcast "$(get_time) Saving done"
+    rconcli broadcast "$(get_time) Creating backup..."
 }
 
 function broadcast_backup_success() {
-    rconcli 'broadcast Backup-done'
+    rconcli broadcast "$(get_time) Backup done"
 }
 
 function broadcast_backup_failed() {
-    rconcli 'broadcast Backup-failed'
+    rconcli broadcast "$(get_time) Backup failed"
 }
 
 function broadcast_player_join() {
-    time=$(date '+%H:%M:%S')
-    rconcli "broadcast ${time}-$1-joined-the-server"
+    rconcli broadcast "$(get_time) $1 joined the server"
 }
 
 function broadcast_player_name_change() {
-    time=$(date '+%H:%M:%S')
-    rconcli "broadcast ${time}-$1-renamed-to-$2"
+    rconcli broadcast "$(get_time) $1 renamed to $2"
 }
 
 function broadcast_player_leave() {
-    time=$(date '+%H:%M:%S')
-    rconcli "broadcast ${time}-$1-left-the-server"
+    rconcli broadcast "$(get_time) $1 left the server"
 }
 
 function check_is_server_empty() {