+++ includes/server.sh
@@ -30,8 +30,8 @@ function stop_server() {
     if [[ -n $RCON_ENABLED ]] && [[ $RCON_ENABLED == "true" ]]; then
         save_and_shutdown_server
     fi
-	kill -SIGTERM "$(pidof PalServer-Linux-Test)"
-	tail --pid="$(pidof PalServer-Linux-Test)" -f 2>/dev/null
+	kill -SIGTERM "$(pidof PalServer-Linux-Shipping)"
+	tail --pid="$(pidof PalServer-Linux-Shipping)" -f 2>/dev/null
     if [[ -n $WEBHOOK_ENABLED ]] && [[ $WEBHOOK_ENABLED == "true" ]]; then
         send_stop_notification
     fi