+++ scripts/restart.sh
@@ -7,6 +7,10 @@ source /includes/colors.sh
 source /includes/server.sh
 source /includes/webhook.sh
 
+function get_time() {
+    date '+[%H:%M:%S]'
+}
+
 function schedule_restart() {
     ew ">>> Automatic restart was triggered..."
     PLAYER_DETECTION_PID=$(<PLAYER_DETECTION.PID)
@@ -22,23 +26,22 @@ function schedule_restart() {
                     send_restart_now_notification
                 fi
                 break
-            else 
+            else
                 ew ">>> Server has still players"
             fi
-            time=$(date '+%H:%M:%S')
-            rconcli "broadcast ${time}-AUTOMATIC-RESTART-IN-$counter-MINUTES"
+            rconcli broadcast "$(get_time) AUTOMATIC RESTART IN $counter MINUTES"
         fi
         if [[ -n $RESTART_DEBUG_OVERRIDE ]] && [[ $RESTART_DEBUG_OVERRIDE == "true" ]]; then
             sleep 1
         else
             sleep 60
-        fi 
+        fi
     done
 
     if [[ -n $RCON_ENABLED ]] && [[ $RCON_ENABLED == "true" ]]; then
-        rconcli 'broadcast Saving-world-before-restart...'
-        rconcli 'save'
-        rconcli 'broadcast Saving-done'
+        rconcli broadcast "$(get_time) Saving world before restart..."
+        rconcli save
+        rconcli broadcast "$(get_time) Saving done"
         sleep 15
         kill -SIGTERM "${PLAYER_DETECTION_PID}"
         rcon "Shutdown 10"
@@ -50,8 +53,8 @@ function schedule_restart() {
         if [[ -n $WEBHOOK_ENABLED ]] && [[ $WEBHOOK_ENABLED == "true" ]]; then
             send_stop_notification
         fi
-        kill -SIGTERM "$(pidof PalServer-Linux-Test)"
-        tail --pid="$(pidof PalServer-Linux-Test)" -f 2>/dev/null
+        kill -SIGTERM "$(pidof PalServer-Linux-Shipping)"
+        tail --pid="$(pidof PalServer-Linux-Shipping)" -f 2>/dev/null
         kill -SIGTERM "${PLAYER_DETECTION_PID}"
         ew ">>> Server stopped gracefully"
         exit 143;